Audit fixes: models, auth, frontend, SQL
Multiple audit-driven corrections: migrated many Integer columns to BigInteger across models to avoid overflow; added admin verification (Depends(get_current_user)) to plan and mockoon routers and fixed mockoon endpoint typo; removed debug prints and token leak prints; fixed plan ID race by using func.max(); standardized frontend API base to "/api" and migrated register inline JS into controllers (improved UI behavior, card brand detection, typewriter, dev presets, multi-step register flow); removed inline CSS from user settings and referenced stylesheet; cleaned up SQL script (removed DROP block and decorative lines) and added docs/auditoria.md summarizing changes.

+++ backend/models/kpis_maestro_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Numeric
+from sqlalchemy import Column, BigInteger, String, Numeric
 
 from models.base import Base
 
@@ -7,7 +7,7 @@ class KpiMaestro(Base):
 
     __tablename__ = "kpis_maestro"
 
-    id_kpi = Column(Integer, primary_key=True, autoincrement=True)
+    id_kpi = Column(BigInteger, primary_key=True, autoincrement=True)
 
     dimension = Column(String(50), nullable=False)
