+++ backend/models/rendimiento_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, DateTime, Date, ForeignKey
+from sqlalchemy import Column, BigInteger, DateTime, Date, ForeignKey
 
 from models.base import Base
 
@@ -8,9 +8,9 @@ class Rendimiento(Base):
 
     __tablename__ = "rendimiento_publicacion"
 
-    id_rendimiento_publi = Column(Integer, primary_key=True, autoincrement=True)
+    id_rendimiento_publi = Column(BigInteger, primary_key=True, autoincrement=True)
 
-    id_publicacion = Column(Integer, ForeignKey("publicacion.id_publicacion", ondelete="CASCADE"), nullable=False)
+    id_publicacion = Column(BigInteger, ForeignKey("publicacion.id_publicacion", ondelete="CASCADE"), nullable=False)
 
     fecha_captura = Column(DateTime(timezone=True), nullable=False, server_default="CURRENT_TIMESTAMP")
 
@@ -18,6 +18,6 @@ class Rendimiento(Base):
 
     fecha_fin_periodo = Column(Date, nullable=False)
 
-    visitas = Column(Integer, nullable=False, default=0)
+    visitas = Column(BigInteger, nullable=False, default=0)
 
-    ventas = Column(Integer, nullable=False, default=0)
+    ventas = Column(BigInteger, nullable=False, default=0)
